divide tests into different types; organize config files better; configure different types of tests
