+++ .eslintrc.js
@@ -1,22 +1,23 @@
 module.exports = {
-  'extends': 'airbnb',
-  'env': {
-    'node': true,
-    'browser': true,
-    'es6': true
+  extends: 'airbnb',
+  env: {
+    node: true,
+    browser: true,
+    es6: true,
   },
-  'plugins': [
+  plugins: [
     'react', 'import'
   ],
-  'rules':   {
-    'curly': ['error', 'all'],
-    'arrow-parens': ['error', 'always']
+  rules: {
+    curly: ['error', 'all'],
+    'arrow-parens': ['error', 'always'],
+    'space-before-function-paren': ['error', 'never']
   },
-  'parser': 'babel-eslint',
-  'settings': {
+  parser: 'babel-eslint',
+  settings: {
     'import/resolver': {
-      'node': {
-        'moduleDirectory': ['node_modules', 'app']
+      node: {
+        moduleDirectory: ['node_modules', 'app']
       },
     },
   },
