+++ package.json
@@ -15,8 +15,11 @@
     "lint:app": "eslint app",
     "lint:src": "eslint src",
     "lint:test": "eslint test",
-    "test": "yarn test:unit",
-    "test:unit": "NODE_PATH=./app mocha --require babel-register --require test/app/.testSetup.js test/app/*/**.spec.js"
+    "test": "yarn test:unit && yarn test:browser && yarn test:http",
+    "test:browser": "NODE_PATH=./app mocha --require babel-register --require test/app/browser.testSetup.js test/app/*/**.browser.js",
+    "test:http": "NODE_PATH=./app mocha --require babel-register --require test/app/http.testSetup.js test/app/*/**.http.js",
+    "test:unit": "NODE_PATH=./app mocha --require babel-register --require test/app/spec.testSetup.js test/app/*/**.spec.js",
+    "ci": "yarn lint && yarn build && yarn test"
   },
   "author": "Milo Mordaunt <milomord@gmail.com>",
   "license": "MIT",
@@ -35,6 +38,8 @@
     "express": "^4.13.4",
     "express-session": "^1.13.0",
     "hoist-non-react-statics": "^1.0.6",
+    "jsdom": "^9.9.1",
+    "jsdom-global": "^2.1.1",
     "lodash": "^4.13.1",
     "marked": "~0.3.5",
     "mocha": "^3.2.0",
