+++ test/app/browser.testSetup.js
@@ -1,0 +1,8 @@
+import 'jsdom-global/register';
+
+import createLoadTested from './common.testSetup.js';
+
+import chaiEnzyme from 'chai-enzyme';
+global.chai.use(chaiEnzyme());
+
+global.loadTested = createLoadTested('.browser.js');
