+++ test/app/spec.testSetup.js
@@ -1,0 +1,3 @@
+import createLoadTested from './common.testSetup.js';
+
+global.loadTested = createLoadTested('.spec.js');
